Fix error with 'terminate' command for worker
Removed the 'terminate' command for worker when user clicks 'leave'.
The terminate method deletes the web worker completely, which requires
re-creating the web worker when the user joins another project.

While the user's client will still finish the job and send it to the
server, the server reject the job because it does not belong to an
existing worker.