Fixed conflicts in Project
Functions need to be saved to the database as strings, so the
reduceResults property needs to exist as a string in the Project
object. As we're not generating projects on the server side anymore,
this should be okay? If we need to start making server-side projects
again for machine learning, we can figure out a different workaround :+1: